go/pkg/ncraft/kvstore/badgerdb/badgerdb_test.go
package badgerdb

import (
    "context"
    "github.com/stretchr/testify/assert"
    "os"
    "testing"
)

// TestStore tests if reading from, writing to and deleting from the store works properly.
// A struct is used as value. See TestTypes() for a test that is simpler but tests all types.
func TestStore(t *testing.T) {
    // Test with JSON
    store, err := NewStore(DefaultOptions)
    assert.NoError(t, err)

    defer func() {
        os.RemoveAll(DefaultOptions.Dir)
    }()

    store.Put(context.Background(), []byte("test"), []byte("value"))

    value, err := store.Get(context.Background(), []byte("test"))
    assert.Equal(t, value, []byte("value"))

    /*t.Run("Raw", func(t *testing.T) {
    	store, path := createStore(t, encoding.JSON)
    	defer cleanUp(store, path)
    	test.TestStore(store, t)
    })

    // Test with gob
    t.Run("gob", func(t *testing.T) {
    	store, path := createStore(t, encoding.Gob)
    	defer cleanUp(store, path)
    	test.TestStore(store, t)
    })*/
}
